ported progress, scripts & git
