+++ src/git/errors.rs
