+++ src/main.rs
@@ -31,9 +31,13 @@ mod cache;
 mod config;
 mod contexts;
 mod env;
+mod git;
+mod git;
 mod hash;
+mod progress;
 mod proxy;
 mod requests;
+mod scripts;
 mod shell;
 mod system;
 mod version;
