+++ src/scripts/installer.rs
